chore: move out example typed logs into new file
